requirements.txt
sentence-transformers==2.7.0
chromadb==0.4.22
pdfplumber==0.10.3
python-docx==1.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
huggingface-hub==0.23.0
flask==3.0.0
flask-cors==4.0.0